requirements.txt
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==3.7.1
argcomplete==3.5.0
async-generator==1.10
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
autoflake==2.3.1
bcrypt==4.2.0
black==24.8.0
cffi==1.17.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
commitizen==3.29.0
cryptography==43.0.0
decli==0.6.2
distlib==0.3.8
fastapi==0.112.1
fastapi-login==1.10.1
filelock==3.15.4
h11==0.14.0
identify==2.6.0
idna==3.7
iso8601==1.1.0
isort==5.13.2
Jinja2==3.1.4
MarkupSafe==2.1.5
mypy-extensions==1.0.0
nodeenv==1.9.1
outcome==1.3.0.post0
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.2.2
pre-commit==3.8.0
prompt-toolkit==3.0.36
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.2.0
PyJWT==2.9.0
pypika-tortoise==0.1.6
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
questionary==2.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.32
starlette==0.38.2
termcolor==2.4.0
tomlkit==0.13.2
tortoise-orm==0.21.6
trio==0.21.0
typing_extensions==4.12.2
uvicorn==0.30.6
virtualenv==20.26.3
wcwidth==0.2.13
